+++ .github/workflows/code_quality.yml
@@ -256,7 +256,7 @@ jobs:
       - name: PHP Copy Paste Detector
         uses: StephaneBour/actions-php-cpd@7.4
         with:
-          args: src tests --exclude src/Admin --exclude src/Entity  --exclude tests/phpUnit
+          args: src tests --exclude src/Admin --exclude tests/phpUnit
   #
   # phpcf (Php Code Fixer):
   #
