+++ src/Entity/SpecialProgram.php
@@ -1,0 +1,117 @@
+<?php
+
+namespace App\Entity;
+
+use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\HttpFoundation\File\File;
+
+class SpecialProgram
+{
+  public ?File $file = null;
+
+  public ?int $removed_id = null;
+
+  public ?string $old_image_type = null;
+
+  /**
+   * @ORM\Id
+   * @ORM\Column(type="integer")
+   * @ORM\GeneratedValue(strategy="AUTO")
+   */
+  protected ?int $id = null;
+
+  /**
+   * @ORM\Column(type="string")
+   */
+  protected string $imagetype;
+
+  /**
+   * @ORM\Column(type="boolean")
+   */
+  protected bool $active = true;
+
+  /**
+   * @ORM\ManyToOne(targetEntity="Flavor", fetch="EAGER")
+   */
+  protected ?Flavor $flavor = null;
+
+  /**
+   * @ORM\Column(type="integer")
+   */
+  protected int $priority = 0;
+
+  /**
+   * @ORM\Column(type="boolean", options={"default": false})
+   */
+  protected bool $for_ios = false;
+
+  /**
+   * @ORM\ManyToOne(targetEntity="Program", fetch="EAGER")
+   */
+  protected ?Program $program = null;
+
+  public function getFlavor(): ?Flavor
+  {
+    return $this->flavor;
+  }
+
+  public function setFlavor(Flavor $flavor): void
+  {
+    $this->flavor = $flavor;
+  }
+
+  public function getImageType(): string
+  {
+    return $this->imagetype;
+  }
+
+  public function getId(): ?int
+  {
+    return $this->id;
+  }
+
+  public function getProgram(): ?Program
+  {
+    return $this->program;
+  }
+
+  public function getActive(): bool
+  {
+    return $this->active;
+  }
+
+  public function getPriority(): int
+  {
+    return $this->priority;
+  }
+
+  public function setPriority(int $priority): void
+  {
+    $this->priority = $priority;
+  }
+
+  public function getForIos(): bool
+  {
+    return $this->for_ios;
+  }
+
+  public function setForIos(bool $for_ios): void
+  {
+    $this->for_ios = $for_ios;
+  }
+
+  public function isExample(): bool
+  {
+    return true;
+  }
+
+  public function getName(): string
+  {
+    return $this->program->getName();
+  }
+
+  public function getUser(): ?User
+  {
+    return $this->getProgram()->getUser();
+  }
+}
